requirements.txt
Flask==3.0.0
APScheduler==3.10.4
psycopg2-binary==2.9.7